requirements.txt
aioredis==1.3.1
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
Brotli==1.0.9
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dash==1.20.0
dash-bootstrap-components==0.13.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-crispy-forms==1.11.2
django-excel==0.0.10
django-extensions==3.1.3
django-import-export==2.5.0
django-pandas==0.6.4
django-plotly-dash==1.6.4
django-redis==4.12.1
django-session-security==2.6.6
dpd-components==0.1.0
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.0
Flask-Compress==1.9.0
future==0.18.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lml==0.1.0
MarkupPy==1.14
MarkupSafe==2.0.0
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
plotly==4.14.3
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-ods==0.6.0
pyexcel-webio==0.1.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
redis==3.5.3
requests==2.26.0
retrying==1.3.3
Send2Trash==1.5.0
service-identity==21.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
terminado==0.9.5
testpath==0.4.4
texttable==1.6.4
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.4.0
